billy/clean-up: #684 - omoved sections into their own React component to reduce cognitive overload. #685 - added an error handler for when a widget fails to load or when all the widgets fails to load
