+++ src/common/components/YouTubeEmbed.css
@@ -1,0 +1,17 @@
+.video-wrapper {
+  border-radius: 1rem;
+  position: relative;
+  width: 100%;
+  padding-bottom: 56.25%;
+  margin-bottom: 48px;
+  background-color: #000;
+  overflow: hidden;
+}
+.video-wrapper .video-iframe {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  border: 0;
+}
