+++ src/common/home/sections/contributors.css
@@ -1,0 +1,117 @@
+/* Home Contributors */
+.home-contributors {
+  position: relative;
+  padding: 5rem 0 6rem 0;
+  background-color: var(--color-neutral-10);
+}
+
+.home-contributors .title-primary {
+  position: relative;
+  margin: 0 0 3rem 0;
+  font-family: var(--ff-accent);
+  font-size: var(--fs-xxl);
+  text-align: center;
+  color: var(--color-neutral-80);
+  line-height: 1.2;
+}
+
+.home-contributors .title-primary strong {
+  position: relative;
+  font-family: var(--ff-accent);
+  position: relative;
+}
+.home-contributors .title-primary strong span {
+  position: relative;
+  font-family: var(--ff-accent);
+}
+
+.home-contributors .title-primary strong:before {
+  content: "";
+  position: absolute;
+  height: 4px;
+  background-color: var(--color-brand-primary);
+  left: 0;
+  right: 0;
+  bottom: 0.4rem;
+  z-index: 0;
+}
+
+.list-contributors {
+  list-style: none;
+}
+
+.list-contributors {
+  position: relative;
+  list-style: none;
+  margin: 0 auto;
+  padding: 0;
+  width: 80%;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  grid-gap: 3rem;
+  width: 80%;
+  max-width: var(--screen-lg-max);
+}
+
+.list-contributors .contributor-anchor {
+  display: block;
+  width: 160px;
+  height: 160px;
+  border-radius: 160px;
+  overflow: hidden;
+  transition: all 0.16s ease-in-out;
+}
+
+.list-contributors .contributor-anchor:hover,
+.list-contributors .contributor-anchor:focus {
+  box-shadow: 0px -1px 26px 0px rgba(var(--color-brand-primary-rgb), 0.59);
+  -webkit-box-shadow: 0px -1px 26px 0px rgba(var(--color-brand-primary-rgb), 0.59);
+  -moz-box-shadow: 0px -1px 26px 0px rgba(var(--color-brand-primary-rgb), 0.59);
+  transform: scale(1.2);
+}
+
+.list-contributors .contributor-anchor .contributor-thumb {
+  width: 160px;
+  height: auto;
+  filter: grayscale(100%);
+}
+
+.list-contributors .contributor-anchor:hover .contributor-thumb,
+.list-contributors .contributor-anchor:focus .contributor-thumb {
+  filter: grayscale(0%);
+}
+
+@media screen and (max-width: 1024px) {
+  .list-contributors {
+    grid-gap: 2rem;
+  }
+
+  .list-contributors .contributor-anchor {
+    width: 120px;
+    height: 120px;
+    border-radius: 120px;
+  }
+
+  .list-contributors .contributor-anchor .contributor-thumb {
+    width: 120px;
+  }
+}
+
+@media screen and (max-width: 568px) {
+  .list-contributors {
+    width: 100%;
+    grid-gap: 1rem;
+  }
+
+  .list-contributors .contributor-anchor {
+    width: 80px;
+    height: 80px;
+    border-radius: 80px;
+  }
+
+  .list-contributors .contributor-anchor .contributor-thumb {
+    width: 80px;
+  }
+}
