+++ src/common/home/sections/playground.css
@@ -1,0 +1,39 @@
+/* Home Plays */
+.home-plays {
+  position: relative;
+  padding: 5rem 0 6rem 0;
+  background: rgba(var(--color-brand-primary-rgb), 0.1);
+}
+
+.home-plays:before {
+  content: " ";
+  display: block;
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  opacity: 0.5;
+  background-image: url("../../../images/bg-spikes.webp");
+  mix-blend-mode: overlay;
+}
+
+.home-plays .plays-title-primary {
+  position: relative;
+  margin: 0 0 3rem 0;
+  font-family: var(--ff-accent);
+  font-size: var(--fs-xxl);
+  text-align: center;
+  color: var(--color-neutral-80);
+}
+
+.home-plays .plays-title-primary strong {
+  font-family: var(--ff-accent);
+  position: relative;
+  color: var(--color-neutral-90);
+}
+
+.home-plays .home-plays-footer {
+  margin: 4rem 0 0 0;
+  text-align: center;
+}
